[BZ #2182]
	* math/s_cacosh.c: Return values from positive branch.
	* math/s_cacoshf.c: Likewise.
	* math/s_cacoshl.c: Likewise.

	* sysvipc/msgrcv.c: Likewise.
	* include/sys/msg.h: Likewise.

+++ include/sys/msg.h
@@ -1,8 +1,8 @@
 #ifndef _SYS_MSG_H
 #include <sysvipc/sys/msg.h>
 
-extern int __libc_msgrcv (int msqid, void *msgp, size_t msgsz, long int msgtyp,
-			  int msgflg);
+extern ssize_t __libc_msgrcv (int msqid, void *msgp, size_t msgsz,
+			      long int msgtyp, int msgflg);
 extern int __libc_msgsnd (int msqid, const void *msgp, size_t msgsz,
 			  int msgflg);
